requirements.txt
async-timeout==4.0.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
Deprecated==1.2.13
firebase-admin==6.0.1
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
